Change following PR feedbacks
