+++ tests/Integration/data/managers.php
@@ -2,8 +2,8 @@
 
 namespace Managers;
 
-use App\DummyManagerReturnsContract;
-use App\DummyManagerReturnsImpl;
+use Illuminate\Support\Manager;
+
 
 function testManagerReturnsContract(DummyManagerReturnsContract $manager): void {
     $manager->type();
@@ -13,4 +13,43 @@ function testManagerReturnsContract(DummyManagerReturnsContract $manager): void 
 function testManagerReturnsImplementation(DummyManagerReturnsImpl $manager): void {
     $manager->type();
     $manager->publicMethodNotInContract();
+}
+
+
+class DummyDriver implements DummyDriverContract
+{
+    public function type(): string {
+        return 'dummy';
+    }
+
+    public function publicMethodNotInContract(): string {
+        return 'Hello World';
+    }
+}
+
+interface DummyDriverContract
+{
+    public function type(): string;
+}
+
+class DummyManagerReturnsContract extends Manager
+{
+    public function getDefaultDriver(): string {
+        return 'dummy';
+    }
+
+    public function createDummyDriver(): DummyDriverContract {
+        return app(DummyDriver::class);
+    }
+}
+
+class DummyManagerReturnsImpl extends Manager
+{
+    public function getDefaultDriver(): string {
+        return 'dummy';
+    }
+
+    public function createDummyDriver(): DummyDriver {
+        return app(DummyDriver::class);
+    }
 }
