--- tests/application/app/DummyDriver.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace App;
-
-
-class DummyDriver implements DummyDriverContract
-{
-    public function type(): string {
-        return 'dummy';
-    }
-
-    public function publicMethodNotInContract(): string {
-        return 'Hello World';
-    }
-}
