--- tests/application/app/DummyDriverContract.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace App;
-
-
-interface DummyDriverContract
-{
-    public function type(): string;
-}
