--- tests/application/app/DummyManagerReturnsContract.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Support\Manager;
-
-/**
- * @mixin DummyDriverContract
- */
-class DummyManagerReturnsContract extends Manager
-{
-    public function getDefaultDriver(): string {
-        return 'dummy';
-    }
-
-    public function createDummyDriver(): DummyDriverContract {
-        return app(DummyDriver::class);
-    }
-}
